feat : 유니버스 조회 기능 구현

+++ src/main/java/com/hoo/aoo/admin/domain/universe/SocialInfo.java
@@ -6,11 +6,11 @@ import java.util.List;
 
 @Getter
 public class SocialInfo {
-    private final Long likeCount;
+    private final Integer likeCount;
     private final Long viewCount;
     private final List<String> hashtags;
 
-    public SocialInfo(Long likeCount, Long viewCount, List<String> hashtags) {
+    public SocialInfo(Integer likeCount, Long viewCount, List<String> hashtags) {
         this.likeCount = likeCount;
         this.viewCount = viewCount;
         this.hashtags = hashtags;
